chore: fix circular dep
❯ npx madge --circular --extensions ts,js ./src
Processed 243 files (701ms) (45 warnings)

✖ Found 10 circular dependencies!

1) redux/reducers/service-account-reducer.ts >
   smart-components/group/service-account/add-group-service-accounts.tsx
2) redux/reducers/service-account-reducer.ts >
   smart-components/group/service-account/add-group-service-accounts.tsx >
   smart-components/group/add-group/service-accounts-list.tsx
3) smart-components/group/service-account/add-group-service-accounts.tsx >
   smart-components/group/add-group/service-accounts-list.tsx
4) smart-components/group/add-group/add-group-wizard.js >
   smart-components/group/add-group/schema.js
5) smart-components/group/add-group/add-group-wizard.js >
   smart-components/group/add-group/schema.js >
   smart-components/group/add-group/review-template.js
6) smart-components/role/add-role-permissions/add-role-permission-wizard.js >
   smart-components/role/add-role-permissions/schema.js >
   smart-components/role/add-role-permissions/review-template.js
7) smart-components/role/add-role/add-role-success.js >
   smart-components/role/add-role/add-role-wizard.js
8) smart-components/role/add-role/add-role-wizard.js >
   smart-components/role/add-role/schema.js
9) smart-components/role/add-role/add-role-wizard.js >
   smart-components/role/add-role/schema.js >
   smart-components/role/add-role/review-template.js
10) App.js >
    redux/actions/group-actions.js >
    AppEntry.js
