+++ package.json
@@ -115,6 +115,7 @@
     "identity-obj-proxy": "^3.0.0",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
+    "madge": "^8.0.0",
     "npm-run-all": "^4.1.5",
     "prop-types": "^15.8.1",
     "resolve-url-loader": "^5.0.0",
@@ -144,8 +145,9 @@
     "translations:extract": "npx formatjs extract ./src/*.js --out-file ./locales/translation-template.json",
     "translations:compile": "npx formatjs compile ./locales/translation-template.json --out-file ./src/locales/translations.json",
     "translations:datafile": "node scripts/createDataJson.js",
-    "deploy": "npm-run-all build lint test",
-    "verify": "npm-run-all build lint test",
+    "circular-dependencies": "madge --circular --extensions ts,js ./src",
+    "deploy": "npm-run-all build circular-dependencies lint test",
+    "verify": "npm-run-all build circular-dependencies lint test",
     "nightly": "npm run deploy",
     "postinstall": "ts-patch install && rm -rf .cache"
   },
