+++ src/locales/locale.ts
@@ -1,0 +1,1 @@
+export const locale = 'en';
