+++ src/smart-components/role/add-role-permissions/add-role-permission-wizard-context.js
@@ -1,0 +1,7 @@
+import { createContext } from 'react';
+
+export const AddRolePermissionWizardContext = createContext({
+  success: false,
+  submitting: false,
+  error: undefined,
+});
