+++ src/smart-components/role/add-role/add-role-wizard-context.js
@@ -1,0 +1,7 @@
+import { createContext } from 'react';
+
+export const AddRoleWizardContext = createContext({
+  success: false,
+  submitting: false,
+  error: undefined,
+});
